Implement the class SimpleContext
Add the SimpleContext class which is an implementation of
the interface Context. A context is use by several object
of the simulation to store data which can be shared with
external components which maybe want treat these data
to for example display it or just make computation. Thanks
for context we can make the simulation more user-friendly
and user can more easily study the simulation result.