Implement Environment and PhysicalNetwork classes
Implement Environment class which represents an Environment where
SimpleAgent evolve in.
Also implement PhysicalNetwork class which allows the simulation
of physical connection between SimpleAgent and manage how Event
are sent to SimpleAgent. PhysicalNetwork works with PhysicalEvent.
When a PhysicalNetwork must send an Event, it will encapsulate it
in a PhysicalEvent. This PhysicalEvent must after be manage by a
Protocol which is able to process this type of Event.

Review the package structure. Now Protocol and Behavior packages
are sub package of the Agent package. It is more logic because
Behavior and Protocol are highly associated to Agent and only
Agent must use Behavior and Protocol.