#25zqjtr[in progress] Implement Scheduler and Executable interface.
Define the Scheduler interface and Executable interface. A Scheduler schedules
and execute Executables. An Executable execution can be stopped with the
waiting mechanism of Scheduler by calling the methods Scheduler#await. Executable
wait on Condition which decide when the execution must be resumed. The execution
is resumed when the method Condition#wakeup is called.

Update the .gitlab-ci.yml to remove artifact from job which does not generate it
and now the Sonarqube job is also done on all event on Master and Develop branches