#25zqjtr[in progress] Make an implementation of Executor: MultiThreadExecutor
MultiThreadExecutor is an implementation of the interface Executor. As its name says, it is
based on multi thread implementation. When you create a MultiThreadExecutor, you define a max
number of thread which is the max number of thread which will concurrently executing. However,
it is possible that more than this number of thread are created.

Make an implementation of MultiThreadExecutor which appears not bugged. This implementation
try to reduce the number of synchronisation. Only one lock is use (a ReentrantLock) and several
Condition are generated from this lock. The lock is use only if needed, the code is more clear.

This implementation can support single thread execution.You juste need to specified as 1 the
max number of thread.