#26b558d[in progress] Debug MultiThreadExecutor
Fix the MultiThreadExecutor. When the Executor was shutdown, waiting
ExecutorThread are Interrupted and enter the catch block. However,
the activeThreads counter was not increase after their "forced wake
up" but it was decrease after the end of the ExecutorThread execution (see ExecutorThread#execute) and the counter becomes negative at the Executor
termination. Now, when the ExecutorThread is interrupted, it increases the
active Thread counter in the catch block, therefore the counter cannot
be negative anymore and Scheduler#isTerminated() can returns true correctly.