#26b9fax[in progress] Debug MultiThreadExecutor.InternalThread wakeup await bug
Fix the MultiThreadExecutor.InternalThread wakeup await bug. It has always
been that if a Thread wakeup to early another Thread which will call wait but
the wakeup is done before the wait, then the Thread will never been wake up, but
it is normal. The bug was due to the fact that when a Thread call the method
wakeup, it increases the active thread counter and for some cases, this counter
can became incoherent.

To solve this problem, the method MultiThreadExecutor.InternalThread#increaseActiveThreads
also increase an internal counter which counts the number of times the increase
methods has been called. When the method MultiThreadExecutor.InternalThread#decreaseActiveThreads,
it decreases the active thread counter the number of internal counter.
In plus of that, for both methods wakeup and await, the increase and decrease methods
are called in the synchronized(getLockMonitor()) block.