#25zqjvj[in progress] Create PalmBeachSimulation and all around classes.
Create the PalmBeachSimulation which is the class which "manage" the Simulation. A JVM
can only have one instance of a PalmBeachSimulation. To access to simulation features,
the simulation must be started and use PalmBeachSimulation static methods.

Create the SimulationConfiguration which read a simulation configuration file and
create a PalmBeachSimulation base on it. An example of a complete simulation
configuration can be found in the file simulation.conf. The configuration library
used to facilite the read of this type of file is TypeSafe.config.

Update SimpleAgent, now Agent can evolve in several Environment and SimpleAgent
does not have a reference on which Environments it is evolving. To know that, it
must ask the simulation the list of Environment and call for each the method
Environment#agentIsEvolving. It is assumed that it does not cost many ressources
because there are not many Environment in a Simulation.