#25zqjvj[in progress] UT for all Configuration classes and Simulation classes
All UT have been done for Configuration classes.
All UT have been done for Simulation classes.
The code coverage cannot be 100% with UT, but it can be increased with Integration
tests.

At this point, Simulation can be configured, created and started, However some
modifications must be done to be really "user-friendly" and usable.