#27qpjax[in progress] Update Environment and PhysicalNetwork
Now PhysicalNetwork constructor need now an Environment. A PhysicalNetwork is always
associated to an Environment.

Environment can now accept observer. Environment.EnvironmentObserver are notified
when a SimpleAgent is added or removed.

When a PhysicalNetwork is created with an Environment, it automatically adds him
to the Environment as Environment.EnvironmentObserver. In that way, PhysicalNetwork
can update itself in function of the add and remove of SimpleAgent of the Environment
with which it is associated.

All UT pass, Environment package is always covered at 100%