#27xe46b[in progress] Remove PhysicalNetwork and change it to Network
Conception error with PhysicalNetwork. A better abstraction is to do a class
Network which represent Agent Network and with which we can verify if
two Agents have connection between them or not. In that way, there is no
physical simulation, but it is not needed for the moment and this layer can
be done latter over the Network class.