#27xfrv8[in progress] Create Messaging classes
Create the interface Messenger which is a combination of two interface MessageSender
and MessageReceiver. A Messenger can send and received. Because Messenger received
message, it is possible to read Messages that it has received.

MessageProtocol is just Protocol which just inherits of MessageReceiver. This abstract
class manage how to receive and deliver Message, but also how to read received Messages.
From this class, it is possible to wait until a message is received.

Create the class SimpleMessageSender which just try to send Message across a Network.

Create the class SimpleBroadcast which just try to broadcast Message accros a Network.