#27ckpe1[in progress] Add the possibility to make dependencies between Protocol in configuration file
It is now possible to specify dependencies between Protocol in configuration file. Sometimes, Protocol
need to use others "sub" Protocols to work. To facilitate the implementation of the Simulation, it is
now possible to specify in the configuration file dependencies between Protocol.

How it works:
- First in the implementation of your Protocol, you must define several fields which are the "sub"
Protocols that you need to use. Each of these fields must at least have a setter (with standard name)
and must be NOT final
- Second, in the configuration file, you define your protocol has before, and you can define protocol
dependencies like this: protocol.yourProtocol.myDependencyName=protocolIdentifier

In that way when the PalmBeachSimulation is created, the set of the protocol dependency is automatically
done.