#2838pp2[in progress] Review Messaging architecture
Review the Messaging architecture to be more general and flexible.

Now MessageReceiver can have MessageReceiver.MessageReceiverObserver which are notified
when the MessageReceiver deliver a Message.Therefore, it is now possible to make "layers"
of MessageReceiver which observe each others.

SimpleBroadcast use a Messenger to send its messages. Therefore, SimpleBroadcast observer
the message reception of its Messenger to trigger its intern reception