#283ce60[in progress] Add AgentProcessEventExecutable
AgentProcessEventExecutable is an Executable which just call
the method SimpleAgent.processEvent. The particularity of this
Executable it returns the agent instance with the methode
getLockMonitor(). In that way, the user does not need to have
conscience of the synchronization.

Update versioning in pom.xml

Update .gitlab-ci.yml to allow deployment of snapshots.