#288yqdx[in progress] Create BestEffortBroadcast and ReliableBroadcast
BestEffortBroadcast is a broadcast protocol which try to send message to all agents
in the Network but with no guaranty that the message will be received.

ReliableBroadcast is a broadcast protocol which guaranty that if a correct agent
deliver a message, therefore all correct agents finally deliver this message.

Update Network.send() which now verify if the source agent is started. In that way, it
is possible to simulate broadcast fail by stopped an agent when it is broadcasting a
message. Update UT in consequences.

Update PalmBeachRunner which now does not start if there already is a PalmBeachSimulation
singleton set. Update UT in consequences.

Update pom.xml version.