#2u43x8e[in progress] Create PulsingMessenger
PulsingMessenger is a Messenger which can send message to a receiver
which is not directly connected to the sender. For example, PulsingMessenger
guaranties that the message will be received in a RandomConnectedNetwork. This
is not the case for a SimpleMessenger

Update Network by renaming the method Network.directNeighbors()

Update pom.xml