#28u80hp[in progress] Create Blockchain module*
Create the Blockchain module which contains several objects which
represents the data structure of a Blockchain. However, these new
objects are here to simplify the implementation of an implementation
of a SIMULATION of a Blockchain, not a real implementation of a
Blockchain.

NonForkBlockchain is a Blockchain structure which does not accept
fork. There is only one chain maintain in memory.

MoneyTx is a Transaction that can be a representation of a money
Transaction.