Fix bug for SimulationFinisher call
The Simulation finisher is called when the Simulation end via
a SchedulerObserver. However, SchedulerObserver methods the most
of the time are called by the StepWatcher Thread of the SimpleScheduler.
Before, the StepWatcher Thread was set to daemon, it is not the case
anymore to allow the StepFinisher to execute correctly before the
Main Thread die.